backend/requirements.txt
Flask
flask-cors
ibm-db
teradatasql
python-dotenv
